card_generation/requirements.txt
GitPython == 3.1.*
